Add installation instructions
To add a new page to the documentation, you create an additional .rst
file with the contents, and then add a link to this file in the
`index.rst` page. Note how we had to add two colons `::` to the end of a
line in order to format the next line as code.