Adding API documentation
This patch uses the extension `sphinx-autodoc` to automatically extract
API documentation for the entire package from docstrings. To do this, we
add autodoc to the list of extensions in the Sphinx `conf.py` file, and
add an extra command to the make invocation that runs the script
`sphinx-autodoc` before building the documentation.